python/requirements.txt
pydbus >= 0.6.0
